marmite: fix inverted microphone channels for camcorder
* SND_DEVICE_IN_HANDSET_STEREO_DMIC is used if CAMCORDER is used
as the source, or any other device unassociated with voice calls or VOIP
communication.

We use stereo for the camcorder, so each audio channel has its own microphone. In
the current configuration, the microphone at the bottom of the device (the main
microphone) corresponds to the left audio channel, and the microphone at the top
corresponds to the right audio channel. This setup is inverted.

Swap the microphones used for each channel to fix this.

Change-Id: I4bd0270f72fc1e942f3b7decf625fdb7e93feb3a